11/5/2025 Commit - train code
Did a TON of debugging for running the training code. Most of it was because I used the wenet instead of the interspeech file. Patched all deepspeed and SOX stuff.

Training seeeeems to work now, except now the data is throwing an error, so committing a quick progress thing before I go and question everything I did with the data.